Changed NeighborhoodRadius type from double to int.

+++ MeanImageFilter/vtkMeanImageFilter.h
@@ -32,17 +32,14 @@ public:
   vtkTypeMacro(vtkMeanImageFilter, vtkImageAlgorithm);
   void PrintSelf(ostream& os, vtkIndent indent);
 
-  //void SetStandardDeviation(double sigma0, double sigma1, double sigma2);
-  //void SetStandardDeviation(double sigma[3]);
-
-  vtkSetVector3Macro(NeighborhoodRadius, double);
-  vtkGetVector3Macro(NeighborhoodRadius, double);
+  vtkSetVector3Macro(NeighborhoodRadius, int);
+  vtkGetVector3Macro(NeighborhoodRadius, int);
 
 protected:
   vtkMeanImageFilter();
   ~vtkMeanImageFilter();
 
-  double NeighborhoodRadius[3];
+  int NeighborhoodRadius[3];
 
   vtkImageExport*                VTKExporter;
   //BTX
